Add batch inference time metrics

+++ src/hsi_compression/engine/validate.py
@@ -68,7 +68,7 @@ def validate_one_epoch(
             dtype=torch.float16 if device.type == "cuda" else torch.bfloat16,
         ):
             outputs = model(x)
-            x_hat = outputs["x_hat"]
+            x_hat = outputs["x_hat"].float()
             z = outputs.get("z")
             loss_val = loss_fn(x_hat, x, mask)
